Added combit List & Label 23 Sample to use its .NET component in Java with Javonet
This example demonstrates the usage of the List & Label .NET component (https://www.combit.net/en/reporting-tool) in Java with the third party component Javonet (https://www.javonet.com). This is a strongly typed wrapper for the combit List & Label .NET component. It exposes only the key operations required for this sample and can be expanded.